requirements.txt
# Core dependencies
pytest==8.3.3
playwright==1.55.0
pytest-playwright

# Pytest extensions
pytest-ordering
pytest-xdist
pytest-html
pytest-metadata
pytest-asyncio>=1.3.0
git+https://github.com/iloveitaly/pytest-playwright-visual-snapshot.git@v0.3.0

# Reports and utilities
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest-json-ctrf
python-dotenv==1.0.1
requests>=2.32.5
colorama==0.4.6

# Code quality tools
isort
